Now replacing function names
In this commit, I have added the functionality which handles function detection by checking for an opening round bracket somewhere after the occurrence of a data type. After storing the function name by taking a substring, I map it to a dummy name. I am working under the assumption that variable names and function names are mutually exclusive i.e. the name of one is not found as part of the other.